refactor: split up insert_data into players and team specific modules to shorten length
This is not its final state as I think we can optimize this a bit better, but a starting point at least from the 3k file length we had originally
