--- scrape_and_score/db/__init__.py
@@ -1,3 +1,0 @@
-from .util import is_team_empty, is_player_empty
-from .insert_data import insert_teams, insert_players
-from .fetch_data import fetch_all_teams, fetch_all_players
