+++ scrape_and_score/db/insert/utils.py
@@ -1,0 +1,3 @@
+def convert_time_to_seconds(time_str):
+    minutes, seconds = map(int, time_str.split(":"))
+    return float(minutes * 60 + seconds)
